# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BBox": {
        "description": "Axis-aligned bounding box in WGS84.",
        "properties": {
          "max_lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Max Lat",
            "type": "number"
          },
          "max_lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Max Lon",
            "type": "number"
          },
          "min_lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Min Lat",
            "type": "number"
          },
          "min_lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Min Lon",
            "type": "number"
          }
        },
        "required": [
          "min_lon",
          "min_lat",
          "max_lon",
          "max_lat"
        ],
        "title": "BBox",
        "type": "object"
      },
...
      "GeoPoint": {
        "description": "WGS84 longitude/latitude pair. All ASTRA geometry is EPSG:4326 on the wire.",
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "lon",
          "lat"
        ],
        "title": "GeoPoint",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Scenario": {
        "description": "A named, versioned analysis context. Scenarios are diffable objects.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_baseline": {
            "default": false,
            "title": "Is Baseline",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "perturbations": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Named perturbations applied against the baseline scenario.",
            "title": "Perturbations",
            "type": "object"
          },
          "study_area_id": {
            "title": "Study Area Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "study_area_id",
          "created_at",
          "disclaimer"
        ],
        "title": "Scenario",
        "type": "object"
      },
...
      "ScenarioResponse": {
        "description": "A scenario with the study area it is framed by.",
        "properties": {
          "scenario": {
            "$ref": "#/components/schemas/Scenario"
          },
          "study_area": {
            "$ref": "#/components/schemas/StudyArea"
          }
        },
        "required": [
          "scenario",
          "study_area"
        ],
        "title": "ScenarioResponse",
        "type": "object"
      },
...
      "StudyArea": {
        "description": "The geographic frame every layer and fixture must fall inside.",
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BBox"
          },
          "centre": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "default_zoom": {
            "default": 11.0,
            "title": "Default Zoom",
            "type": "number"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "district": {
            "title": "District",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "district",
          "state",
          "bbox",
          "centre",
          "description"
        ],
        "title": "StudyArea",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nASTRA - Proactive Settlement Risk & Relocation Intelligence.\n\nEvery red-zone boundary, capacity figure and priority ranking in ASTRA is produced by a documented, auditable formula or a constraint solver. AI is used only to explain those outputs in plain language, never to produce them.\n\nDecision-support output. Final relocation decisions rest with the SDMA /\nDistrict Authority. Habitation and candidate-site records in the demonstration\nscenario are synthetic and terrain-calibrated, and are not an official hazard\ndesignation of any real settlement.\n",
    "title": "ASTRA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/scenarios/{scenario_id}": {
      "get": {
        "operationId": "scenario_detail_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Detail",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}